+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 > Due to a versioning mistake, the official stable API will be defined in v2.0.0.
 **All `v1.x.x` versions should be considered unstable.**
 
+## Unreleased
+### Added
+- Root API extensions via the `config.extend.root` object.
+
 ## v1.4.0
 ### Added
 - New `before.write` pipeline allows hooks to override values before they reach storage plugins.